refactor: Extract `config` to separate module.
- Use single `logger` named `dvc_studio_client` for all modules.
