+++ tests/test_post_live_metrics.py
@@ -1,12 +1,10 @@
 import logging
-import os
 
 import pytest
-from dulwich.porcelain import clone, init
 from requests import RequestException
 
+from dvc_studio_client import DEFAULT_STUDIO_URL
 from dvc_studio_client.env import (
-    DVC_STUDIO_OFFLINE,
     DVC_STUDIO_REPO_URL,
     DVC_STUDIO_TOKEN,
     DVC_STUDIO_URL,
@@ -14,131 +12,13 @@ from dvc_studio_client.env import (
     STUDIO_TOKEN,
 )
 from dvc_studio_client.post_live_metrics import (
-    STUDIO_URL,
-    _get_remote_url,
-    get_studio_config,
     get_studio_token_and_repo_url,
     post_live_metrics,
 )
 
 
-def test_get_url(monkeypatch, tmp_path_factory):
-    source = os.fspath(tmp_path_factory.mktemp("source"))
-    target = os.fspath(tmp_path_factory.mktemp("target"))
-    with init(source), clone(source, target):
-        monkeypatch.chdir(target)
-        assert _get_remote_url() == source
-
-
-@pytest.mark.parametrize(
-    "token,repo_url",
-    [(DVC_STUDIO_TOKEN, DVC_STUDIO_REPO_URL), (STUDIO_TOKEN, STUDIO_REPO_URL)],
-)
-def test_studio_config_envvar(monkeypatch, token, repo_url):
-    monkeypatch.setenv(token, "FOO_TOKEN")
-    monkeypatch.setenv(repo_url, "FOO_REPO_URL")
-    assert get_studio_config() == {
-        "token": "FOO_TOKEN",
-        "repo_url": "FOO_REPO_URL",
-        "url": STUDIO_URL,
-    }
-
-
-def test_studio_config_dvc_studio_config():
-    dvc_studio_config = {
-        "token": "FOO_TOKEN",
-        "repo_url": "FOO_REPO_URL",
-        "url": "FOO_URL",
-    }
-    expected = {
-        "token": "FOO_TOKEN",
-        "repo_url": "FOO_REPO_URL",
-        "url": "FOO_URL",
-    }
-    assert get_studio_config(dvc_studio_config=dvc_studio_config) == expected
-
-
-def test_studio_config_kwarg(monkeypatch):
-    expected = {
-        "token": "FOO_TOKEN",
-        "repo_url": "FOO_REPO_URL",
-        "url": "FOO_URL",
-    }
-    assert (
-        get_studio_config(
-            studio_token="FOO_TOKEN",
-            studio_repo_url="FOO_REPO_URL",
-            studio_url="FOO_URL",
-        )
-        == expected
-    )
-
-
-def test_studio_config_envvar_override(monkeypatch):
-    monkeypatch.setenv(DVC_STUDIO_TOKEN, "FOO_TOKEN")
-    monkeypatch.setenv(DVC_STUDIO_URL, "FOO_URL")
-    monkeypatch.setenv(DVC_STUDIO_REPO_URL, "FOO_REPO_URL")
-    dvc_studio_config = {
-        "token": "BAR_TOKEN",
-        "url": "BAR_URL",
-    }
-    expected = {
-        "token": "FOO_TOKEN",
-        "repo_url": "FOO_REPO_URL",
-        "url": "FOO_URL",
-    }
-    assert get_studio_config(dvc_studio_config=dvc_studio_config) == expected
-
-
-def test_studio_config_kwarg_override(monkeypatch):
-    monkeypatch.setenv(DVC_STUDIO_TOKEN, "FOO_TOKEN")
-    monkeypatch.setenv(DVC_STUDIO_REPO_URL, "FOO_REPO_URL")
-    monkeypatch.setenv(DVC_STUDIO_URL, "FOO_URL")
-    expected = {
-        "token": "BAR_TOKEN",
-        "repo_url": "BAR_REPO_URL",
-        "url": "BAR_URL",
-    }
-    assert (
-        get_studio_config(
-            studio_token="BAR_TOKEN",
-            studio_repo_url="BAR_REPO_URL",
-            studio_url="BAR_URL",
-        )
-        == expected
-    )
-
-
-@pytest.mark.parametrize(
-    "val",
-    ("1", "y", "yes", "true", True, 1),
-)
-def test_studio_config_offline(monkeypatch, val):
-    monkeypatch.setenv(DVC_STUDIO_TOKEN, "FOO_TOKEN")
-    monkeypatch.setenv(DVC_STUDIO_REPO_URL, "FOO_REPO_URL")
-
-    assert get_studio_config() != {}
-
-    assert get_studio_config(offline=val) == {}
-
-    monkeypatch.setenv(DVC_STUDIO_OFFLINE, val)
-    assert get_studio_config() == {}
-
-    monkeypatch.setenv(DVC_STUDIO_OFFLINE, val)
-    assert get_studio_config() == {}
-
-    assert get_studio_config(dvc_studio_config={"offline": True}) == {}
-
-
-def test_studio_config_infer_url(monkeypatch):
-    monkeypatch.setenv(DVC_STUDIO_TOKEN, "FOO_TOKEN")
-    monkeypatch.setenv(DVC_STUDIO_REPO_URL, "FOO_REPO_URL")
-
-    assert get_studio_config()["url"] == STUDIO_URL
-
-
 def test_post_live_metrics_skip_on_missing_token(caplog):
-    with caplog.at_level(logging.DEBUG, logger="dvc_studio_client.post_live_metrics"):
+    with caplog.at_level(logging.DEBUG, logger="dvc_studio_client"):
         assert post_live_metrics("start", "current_rev", "fooname", "fooclient") is None
         assert caplog.records[0].message == (
             "DVC_STUDIO_TOKEN not found. Skipping `post_studio_live_metrics`"
@@ -148,8 +28,8 @@ def test_post_live_metrics_skip_on_missing_token(caplog):
 def test_post_live_metrics_skip_on_schema_error(caplog, monkeypatch):
     monkeypatch.setenv(DVC_STUDIO_TOKEN, "FOO_TOKEN")
     monkeypatch.setenv(DVC_STUDIO_REPO_URL, "FOO_REPO_URL")
-    monkeypatch.setenv(DVC_STUDIO_URL, STUDIO_URL)
-    with caplog.at_level(logging.DEBUG, logger="dvc_studio_client.post_live_metrics"):
+    monkeypatch.setenv(DVC_STUDIO_URL, DEFAULT_STUDIO_URL)
+    with caplog.at_level(logging.DEBUG, logger="dvc_studio_client"):
         assert post_live_metrics("start", "bad_hash", "fooname", "fooclient") is None
         assert caplog.records[0].message == (
             "expected a length 40 commit sha for dictionary value @ "
@@ -443,12 +323,6 @@ def test_post_live_metrics_bad_response(mocker, monkeypatch):
         )
         is False
     )
-
-
-def test_get_studio_config_skip_repo_url(monkeypatch):
-    monkeypatch.setenv(STUDIO_REPO_URL, "FOO_REPO_URL")
-    config = get_studio_config()
-    assert config == {}  # Skipped call to get_repo_url
 
 
 def test_post_live_metrics_token_and_repo_url_args(mocker, monkeypatch):
